Refactored testen.
