+++ config/configurations.py
@@ -10,8 +10,11 @@ IV3_REPO_PATH = "https://raw.github.com/statistiekcbs/iv3_definities/master/"
 IV3_SCHEMA_FILE = "iv3_data_schema_v{}.json"
 IV3_DEF_FILE = "iv3_definities_{}_{}.json"
 
-# Bepaalt of json bestand-controle met definitiebestand extern met API call gebeurt of niet.
-EXTERNE_DEF_BESTAND_CONTROLE = False
+# JSON bestand-controle gebeurt extern via Open State Foundation (OSF) API aanroep of niet.
+EXTERNE_CONTROLE = False
+
+OSF_API_GEEF_TOTALEN_UITKOMST_URL = 'https://cbs.openstate.eu/geef_totalen_uitkomst'
+OSF_API_GEEF_FOUTEN_URL = 'https://cbs.openstate.eu/geef_fouten'
 
 basedir = os.path.abspath(os.path.dirname(__file__))
 load_dotenv(os.path.join(basedir, '.env'))
