--- tests/test_bestanden/voorbeeld_geeft_schema_fouten.json
@@ -1,47 +1,0 @@
-{"metadata":{"overheidslaag":"Gemeente","overheidsnummer":"1234","overheidsnaam":"TheoDorp","boekjaar":"2018","periode":"1","status":"Realisatie","datum":"6-3-2019"},
-"contact":{"naam":"Theo Grivel","telefoon":"070 337 4501","e-mail":"t.grivel@cbs.nl"},
-"data":{
-"lasten":[
-{"categorie":"3.5.1","taakveld":"0.3","bedrag":23},
-{"categorie":"4.3.3","taakveld":"0.5","bedrag":45},
-{"categorie":"4.3.2","taakveld":"1.1","bedrag":5},
-{"categorie":"4.4.8","taakveld":"2.3","bedrag":56},
-{"categorie":"7.4","taakveld":"3.3","bedrag":55},
-{"categorie":"4.3.4","taakveld":"5.2","bedrag":45},
-{"categorie":"4.3.4","taakveld":"6.6","bedrag":11},
-{"categorie":"4.3.9","taakveld":"8.3","bedrag":23}],
-"balans_lasten":[
-{"categorie":"4.4.2","balanscode":"A122","bedrag":54},
-{"categorie":"6.1","balanscode":"A1323","bedrag":4}],
-"baten":[
-{"categorie":"3.5.2","taakveld":"0.1","bedrag":88},
-{"categorie":"4.3.7","taakveld":"0.4","bedrag":61},
-{"categorie":"4.3.7","taakveld":"1.1","bedrag":1},
-{"categorie":"3.6","taakveld":"4.2","bedrag":45},
-{"categorie":"4.4.6","taakveld":"5.4","bedrag":6},
-{"categorie":"7.4","taakveld":"6.6","bedrag":55},
-{"categorie":"3.5.2","taakveld":"7.1","bedrag":53},
-{"categorie":"4.3.4","taakveld":"7.3","bedrag":7}],
-"balans_baten":[
-{"categorie":"4.4.4","balanscode":"A123","bedrag":22},
-{"categorie":"6.1","balanscode":"A1331a","bedrag":45}],
-"balans_standen":[
-{"balanscode":"A122","standper":"1 januari","bedrag":12},
-{"balanscode":"A123","standper":"1 januari","bedrag":50},
-{"balanscode":"A1323","standper":"1 januari","bedrag":20},
-{"balanscode":"A1331a","standper":"1 januari","bedrag":80},
-{"balanscode":"A224","standper":"1 januari","bedrag":80},
-{"balanscode":"A225b","standper":"1 januari","bedrag":20},
-{"balanscode":"P12","standper":"1 januari","bedrag":50},
-{"balanscode":"P136","standper":"1 januari","bedrag":20},
-{"balanscode":"P29b","standper":"1 januari","bedrag":80},
-{"balanscode":"A122","standper":"ultimo","bedrag":66},
-{"balanscode":"A123","standper":"ultimo","bedrag":28},
-{"balanscode":"A1323","standper":"ultimo","bedrag":24},
-{"balanscode":"A1331a","standper":"ultimo","bedrag":35},
-{"balanscode":"A224","standper":"ultimo","bedrag":71},
-{"balanscode":"A225b","standper":"ultimo","bedrag":27},
-{"balanscode":"P12","standper":"ultimo","bedrag":24},
-{"balanscode":"P136","standper":"ultimo","bedrag":38},
-{"balanscode":"P29b","standper":"ultimo","bedrag":24}]
-}}
