+++ tests/test_controles.py
@@ -1,107 +1,10 @@
 from unittest import TestCase
 
-import requests
+from applicatie.logic.controles import geef_codechecklijst
 
 
 class TestControles(TestCase):
-    def test_controle_valide_request_geeft_http500_DIT_MOET_http200_WORDEN(self):
-        api_url = 'https://cbs.openstate.eu/geef_totalen_uitkomst'
-        bestand = "test_bestanden/goed_voorbeeld.json"
-
-        f = open(bestand, 'rb')
-        bestanden = {'file': f}
-        response = requests.post(api_url, files=bestanden)
-
-        f.close()
-
-        self.assertEqual(500, response.status_code, "De HTTP statuscode van de response is niet gelijk aan 500.")
-
-    def test_controle_request_zonder_bestand_geeft_http500(self):
-        api_url = 'https://cbs.openstate.eu/geef_totalen_uitkomst'
-
-        response = requests.post(api_url)
-
-        self.assertEqual(500, response.status_code, "De HTTP statuscode van de response is niet gelijk aan 500.")
-
-    def test_controle_request_met_file_anders_dan_JSON_geeft_http500_en_fouten(self):
-        fouten_element_referentie =\
-            [{'foutcode': 'SF999', 'melding': 'Onjuiste extensie', 'omschrijving': 'er is een algemene fout'}]
-        api_url = 'https://cbs.openstate.eu/geef_totalen_uitkomst'
-        bestand = "test_bestanden/dummy.txt"
-        f = open(bestand, 'rb')
-        bestanden = {'file': f}
-
-        response = requests.post(api_url, files=bestanden)
-        response_json_format = response.json()
-
-        f.close()
-
-        self.assertEqual(500, response.status_code, "De HTTP statuscode van de response is niet gelijk aan 500.")
-
-        fouten_element = response_json_format["fouten"]
-
-        self.assertEqual(fouten_element_referentie, fouten_element,
-                         "Het fouten-element in het bestand bij de response is niet zoals verwacht.")
-
-    def test_controle_request_met_fouten_in_file_geeft_http500_en_fouten(self):
-        fouten_element_referentie = \
-            ["Fout #1: properties, contact, required : 'email' is a required property",
-         'Fout #2: properties, contact, additionalProperties : Additional properties '
-         "are not allowed ('e-mail' was unexpected)",
-         "Fout #3: properties, metadata, required : 'details_openbaar' is a required "
-         'property',
-         "Fout #4: properties, metadata, required : 'financieel_pakket' is a required "
-         'property',
-         "Fout #5: properties, metadata, required : 'export_software' is a required "
-         'property',
-         "Fout #6: properties, metadata, properties, boekjaar, type : '2018' is not of "
-         "type 'number'",
-         "Fout #7: properties, metadata, properties, datum, pattern : '6-3-2019' does "
-         'not match '
-         "'(2[0-3]|[01][0-9]):([0-5][0-9]):([0-5][0-9])(\\\\[0-9]+)?(Z|[+-](?:2[0-3]|[01][0-9]):[0-5][0-9])?$'",
-         "Fout #8: properties, metadata, properties, periode, type : '1' is not of "
-         "type 'number'",
-         "Fout #9: properties, metadata, properties, periode, oneOf : '1' is not valid "
-         'under any of the given schemas',
-         "Subfout #9.1: 0, enum : '1' is not one of [0, 1, 2, 3, 4, 5]",
-         'Samenvatting: totaal 9 unieke fouten gevonden op basis van schema controle.']
-
-        api_url = 'https://cbs.openstate.eu/geef_totalen_uitkomst'
-        bestand = "test_bestanden/voorbeeld_geeft_schema_fouten.json"
-        f = open(bestand, 'rb')
-        bestanden = {'file': f}
-
-        response = requests.post(api_url, files=bestanden)
-        response_json_format = response.json()
-        fouten_element = response_json_format["fouten"]
-
-        f.close()
-
-        self.assertEqual(500, response.status_code, "De HTTP statuscode van de response is niet gelijk aan 500.")
-        self.assertEqual(fouten_element_referentie, fouten_element,
-                         "Het fouten-element in het bestand bij de response is niet zoals verwacht.")
-
-    def test_controle_met_defbestand_extern(self):
-        # json_bestand =
-        #
-        # controle_met_defbestand_extern()
-        #
-        # codefouten = []
-
-        # api_url = 'https://cbs.openstate.eu/geef_totalen_uitkomst'
-        #
-        # response = requests.post(api_url)
-
-        # with open(json_bestand, 'rb') as f:
-
-        # response = requests.post(api_url, files={json_bestand: f})
-
-
-        self.assertTrue(False, "Deze test moet nog geimplementeerd worden.")
-
     def test_geef_codechecklijst(self):
-        from applicatie.logic.controles import geef_codechecklijst
-
         codelijst = {'taakveld': {'codelijst': [{'code': '0.', 'omschrijving': 'BESTUUR EN ONDERSTEUNING',
                                                  'subcodes': [{'code': '0.1', 'omschrijving': 'Bestuur'},
                                                               {'code': '0.10', 'omschrijving': 'Mutaties reserves'},
@@ -618,6 +521,3 @@ class TestControles(TestCase):
                                     'dimensie_2', 'rijkoppen balans_standen tabel', 2, 'dimensies_balans_standen']}
 
         self.assertDictEqual(exp_result, result)
-
-    def test_extract_values_from_json(self):
-        from applicatie.logic.controles import extract_values_from_json
